add threaded killing with limits, remove old code
