--- data/new_monsters/utils/commonLoot.js
@@ -1,37 +1,0 @@
-module.exports = {
-	dragonMedHelm: 1149,
-	dragonHalberd: 3204,
-	pureEssence: 7936,
-	deathRune: 560,
-	chaosRune: 562,
-	lawRune: 563,
-	snapdragon: 3000,
-	dwarfWeed: 267,
-	torstol: 269,
-	toadflax: 2998,
-	coal: 453,
-	dragonstoneBoltTips: 9193,
-	flax: 1779,
-	mahoganyLogs: 6332,
-	runiteOre: 451,
-	snakeskin: 6289,
-	yewLogs: 1515,
-	swampTar: 1939,
-	antidote24: 5952,
-	dragonBones: 536,
-	adamantiteBar: 2361,
-	coconut: 5974,
-	grapes: 1987,
-	zulAndraTeleport: 12938,
-	mantaRay: 391,
-	battlestaff: 1391,
-	spiritSeed: 5317,
-	torstolSeed: 5304,
-	dwarfWeedSeed: 5303,
-	snapdragonSeed: 5300,
-	toadflaxSeed: 5296,
-	palmTreeSeed: 5289,
-	papayaTreeSeed: 5288,
-	calquatTreeSeed: 5290,
-	magicSeed: 5316
-};
