--- data/new_monsters/utils/killWorkerFunction.js
@@ -1,445 +1,0 @@
-const { Monsters } = require('oldschooljs');
-
-const { cleanString } = require('../../../config/util');
-const { addBankToBank } = require('../../../src/lib/util');
-const gauntlet = require('../gauntlet');
-
-const corp = require('../../monsters/corp');
-const raids = require('../../monsters/raids');
-const tob = require('../../monsters/tob');
-const grotesqueGuardians = require('../../monsters/grotesqueGuardians');
-const demonicGorilla = require('../../monsters/demonicGorilla');
-const abbyDemon = require('../../monsters/abbyDemon');
-const kurask = require('../../monsters/kurask');
-const kalphiteQueen = require('../../monsters/kalphiteQueen');
-const lizardmanShaman = require('../../monsters/lizardmanShaman');
-const callisto = require('../../monsters/callisto');
-const vetion = require('../../monsters/vetion');
-const brutalBlackDragon = require('../../monsters/brutalBlackDragon');
-const venenatis = require('../../monsters/venenatis');
-const wyvern = require('../../monsters/wyvern');
-const ancientWyvern = require('../../monsters/ancientWyvern');
-const kingBlackDragon = require('../../monsters/kingBlackDragon');
-const abyssalSire = require('../../monsters/abyssalSire');
-const thermy = require('../../monsters/thermy');
-const skotizo = require('../../monsters/skotizo');
-const scorpia = require('../../monsters/scorpia');
-const chaosElemental = require('../../monsters/chaosElemental');
-const wintertodt = require('../../monsters/wintertodt');
-const kraken = require('../../monsters/kraken');
-const obor = require('../../monsters/obor');
-const lavaDragon = require('../../monsters/lavaDragon');
-const chaosFanatic = require('../../monsters/chaosFanatic');
-const crazyArchaeologist = require('../../monsters/crazyArchaeologist');
-const mithrilDragon = require('../../monsters/mithrilDragon');
-const adamantDragon = require('../../monsters/adamantDragon');
-const runeDragon = require('../../monsters/runeDragon');
-const bryophyta = require('../../monsters/bryophyta');
-const alchemicalHydra = require('../../monsters/alchemicalHydra');
-const hydra = require('../../monsters/hydra');
-const drake = require('../../monsters/drake');
-const wyrm = require('../../monsters/wyrm');
-const hespori = require('../../monsters/hespori');
-const sarachnis = require('../../monsters/sarachnis');
-const basKnight = require('../../../data/monsters/basKnight');
-
-module.exports = function killWorkerFunction(quantity, bossName) {
-	switch (cleanString(bossName)) {
-		case 'SARACHNIS':
-		case 'SARACH':
-		case 'SRARACHA': {
-			if (quantity > 5000) {
-				return 'I can only kill a maximum of 5000 Sarachnis at a time!';
-			}
-			const loot = sarachnis.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'CORP':
-		case 'CORPOREALBEAST':
-		case 'CORPBEAST': {
-			if (quantity > 5000) return 'I can only kill a maximum of 5000 Corp beasts at a time!';
-			const loot = corp.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'KURASK':
-		case 'KURASKS': {
-			if (quantity > 5000) return 'I can only kill a maximum of 5000 Kurasks at a time!';
-			const loot = kurask.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'LAVADRAGON':
-		case 'LAVADRAGONS':
-		case 'LAVADRAGS':
-		case 'LAVADRAG': {
-			if (quantity > 200) return 'I can only kill a maximum of 200 Lava Dragons at a time!';
-			const loot = lavaDragon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'BRUTALBLACKDRAGON':
-		case 'BBD':
-		case 'BRUTALBLACKDRAGONS':
-		case 'BRUTALBLACKDRAGS':
-		case 'BRUTALBLACKS': {
-			if (quantity > 100)
-				return 'I can only kill a maximum of 100 Brutal Black Dragons at a time!';
-			const loot = brutalBlackDragon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'RAIDS':
-		case 'OLM':
-		case 'RAIDS1':
-		case 'COX':
-		case 'CHAMBERSOFXERIC': {
-			if (quantity > 100000) return 'I can only do a maximum of 100,000 Raids at a time!';
-			const loot = raids.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'RAIDS2':
-		case 'THEATREOFBLOOD':
-		case 'TOB': {
-			if (quantity > 10000) {
-				return "I can only do a maximum of 10,000 Theatres of Blood at a time! I'm not Woox!";
-			}
-			const loot = tob.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'BARROWS': {
-			if (quantity > 500) return 'I can only do a maximum of 500 Barrows Chests at a time!';
-			return Monsters.Barrows.kill(quantity);
-		}
-		case 'BANDOS':
-		case 'GENERALGRAARDOR':
-		case 'GRAARDOR': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Bandos at a time!';
-			return Monsters.GeneralGraardor.kill(quantity);
-		}
-		case 'SARA':
-		case 'SARADOMIN':
-		case 'ZILY':
-		case 'ZILYANA':
-		case 'COMMANDERZILYANA': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Sara at a time!';
-			return Monsters.CommanderZilyana.kill(quantity);
-		}
-		case 'MEN':
-		case 'MAN': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Men at a time!';
-			return Monsters.Man.kill(quantity);
-		}
-		case 'WOMEN':
-		case 'WOMAN': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Women at a time! 😉';
-			return Monsters.Woman.kill(quantity);
-		}
-		case 'GOBLIN':
-		case 'GOBLINS': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Goblins at a time!';
-			return Monsters.Goblins.kill(quantity);
-		}
-		case 'GUARD':
-		case 'GUARDS': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Guards at a time!';
-			return Monsters.Guard.kill(quantity);
-		}
-		case 'ZULRAH': {
-			if (quantity > 10000) {
-				return 'I can only do a maximum of 10000 Zulrah kills at a time!';
-			}
-			return Monsters.Zulrah.kill(quantity);
-		}
-		case 'VORKATH': {
-			if (quantity > 500) return 'I can only do a maximum of 500 Vorkath kills at a time!';
-			return Monsters.Vorkath.kill(quantity);
-		}
-		case 'CERB':
-		case 'CERBERUS': {
-			if (quantity > 2000) return 'I can only do a maximum of 2000 Cerberus kills at a time!';
-			return Monsters.Cerberus.kill(quantity);
-		}
-		case 'ABYSSALDEMON':
-		case 'ABYSSALDEMONS':
-		case 'ABBYDEMON':
-		case 'ABBYDEMONS': {
-			if (quantity > 5000)
-				return 'I can only do a maximum of 5000 Abyssal Demon kills at a time!';
-			const loot = abbyDemon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'GORILLAS':
-		case 'DGS':
-		case 'DEMONICGORILLAS':
-		case 'DEMONICGORILLA':
-		case 'DEMONICS': {
-			if (quantity > 500) return "I can only kill 500 Demonic Gorilla's at a time!";
-			const loot = demonicGorilla.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'KQ':
-		case 'KALPHITEQUEEN': {
-			if (quantity > 500) return "I can only kill 500 Kalphite Queen's at a time!";
-			const loot = kalphiteQueen.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'ZAMMY':
-		case 'ZAMORAK':
-		case 'KRIL':
-		case 'KRILTSUTSAROTH': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Kril at a time!';
-			return Monsters.KrilTsutsaroth.kill(quantity);
-		}
-		case 'ARMA':
-		case 'ARMADYL':
-		case 'KREE':
-		case 'KREEARRA': {
-			if (quantity > 5000) return 'I can only do a maximum of 5000 Arma at a time!';
-			return Monsters.Kreearra.kill(quantity);
-		}
-		case 'SHAMANS':
-		case 'LIZARDMANSHAMANS':
-		case 'LIZARDMANSHAMAN':
-		case 'LIZARDMEN': {
-			if (quantity > 1000) return "I can only kill 1000 Lizardman Shaman's at a time!";
-			const loot = lizardmanShaman.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'CALLISTO': {
-			if (quantity > 500) return "I can only kill 500 Callisto's at a time!";
-			const loot = callisto.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'VETION': {
-			if (quantity > 500) return "I can only kill 500 Vet'ions at a time!";
-			const loot = vetion.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'VENENATIS': {
-			if (quantity > 500) return 'I can only kill 500 Venenatis at a time!';
-			const loot = venenatis.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'SKELETALWYVERN':
-		case 'SKELETALWYVERNS':
-		case 'WYVERNS': {
-			if (quantity > 1000) return 'I can only kill 1000 Skeletal Wyverns at a time!';
-			const loot = wyvern.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'ANCIENTWYVERN':
-		case 'ANCIENTWYVERNS':
-		case 'FOSSILISLANDWYVERNS': {
-			if (quantity > 1000) return 'I can only kill 1000 Ancient Wyverns at a time!';
-			const loot = ancientWyvern.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'GROTESQUEGUARDIANS':
-		case 'GGS':
-		case 'DUSK':
-		case 'DAWN': {
-			if (quantity > 500) return 'I can only kill 500 Grotesque Guardians at a time!';
-			const loot = grotesqueGuardians.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'KBD':
-		case 'KINGBLACKDRAGON': {
-			if (quantity > 500) return 'I can only kill 500 King Black Dragons at a time!';
-			const loot = kingBlackDragon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'ABYSSALSIRE':
-		case 'ABBYSIRE':
-		case 'SIRE': {
-			if (quantity > 500) return 'I can only kill 500 Abyssal Sires at a time!';
-			const loot = abyssalSire.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'THERMY':
-		case 'THERMONUCLEARSMOKEDEVIL': {
-			if (quantity > 500) return "I can only kill 500 Thermonuclear smoke devil's at a time!";
-			const loot = thermy.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'GIANTMOLE':
-		case 'MOLE': {
-			if (quantity > 5000) return "I can only kill 5000 Giant Mole's at a time!";
-			return Monsters.GiantMole.kill(quantity);
-		}
-
-		case 'SKOTIZO': {
-			if (quantity > 500) return 'I can only kill 500 Skotizo at a time!';
-			const loot = skotizo.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'SCORPIA': {
-			if (quantity > 500) return 'I can only kill 500 Scorpia at a time!';
-			const loot = scorpia.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'CHAOSELE':
-		case 'CHAOSELEMENTAL': {
-			if (quantity > 500) return 'I can only kill 500 Chaos Elementals at a time!';
-			const loot = chaosElemental.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'WINTERTODT':
-		case 'WINTERTOAD':
-		case 'TODT':
-		case 'WT': {
-			if (quantity > 500) return "I can only kill 500 Wintertodt's at a time!";
-			const loot = wintertodt.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'KRAKEN': {
-			if (quantity > 500) return "I can only kill 500 Kraken's at a time!";
-			const loot = kraken.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'OBOR':
-		case 'HILLGIANTBOSS': {
-			if (quantity > 500) return "I can only kill 500 Obor's at a time!";
-			const loot = obor.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-		case 'REX':
-		case 'DAGANNOTHREX': {
-			if (quantity > 300) return "I can only kill 300 Dagannoth Rex's at a time!";
-			return Monsters.DagannothRex.kill(quantity);
-		}
-		case 'SUPREME':
-		case 'DAGANNOTHSUPREME': {
-			if (quantity > 300) return 'I can only kill 300 Dagannoth Supremes at a time!';
-			return Monsters.DagannothSupreme.kill(quantity);
-		}
-		case 'PRIME':
-		case 'DAGANNOTHPRIME': {
-			if (quantity > 300) return 'I can only kill 300 Dagannoth Primes at a time!';
-			return Monsters.DagannothPrime.kill(quantity);
-		}
-
-		case 'DAGKINGS':
-		case 'DAGANNOTHKINGS':
-		case 'DKS': {
-			if (quantity > 20) return 'I can only kill all the Dagannoth Kings 10 times!';
-			let loot = {};
-			for (let i = 0; i < quantity; i++) {
-				loot = addBankToBank(Monsters.DagannothPrime.kill(), loot);
-				loot = addBankToBank(Monsters.DagannothRex.kill(), loot);
-				loot = addBankToBank(Monsters.DagannothSupreme.kill(), loot);
-			}
-			return loot;
-		}
-		case 'CHAOSFANATIC': {
-			if (quantity > 500) return 'I can only kill 500 Chaos Fanatics at a time!';
-			const loot = chaosFanatic.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'CRAZYARCH':
-		case 'CRAZYARCHAEOLOGIST': {
-			if (quantity > 500) return 'I can only kill 500 Crazy Archaeologists at a time!';
-			const loot = crazyArchaeologist.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'MITHRILDRAGONS':
-		case 'MITHRILDRAGS':
-		case 'MITHDRAGS':
-		case 'MITHDRAGONS': {
-			if (quantity > 1000) return 'I can only kill 1000 Mithril Dragons at a time!';
-			const loot = mithrilDragon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'ADAMANTDRAGONS':
-		case 'ADAMANTDRAGS':
-		case 'ADDYDRAGS':
-		case 'ADDYDRAGONS': {
-			if (quantity > 1000) return 'I can only kill 1000 Adamant Dragons at a time!';
-			const loot = adamantDragon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'RUNEDRAGONS':
-		case 'RUNEDRAGS': {
-			if (quantity > 1000) return 'I can only kill 1000 Rune Dragons at a time!';
-			const loot = runeDragon.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'BRYOPHYTA':
-		case 'MOSSGIANTBOSS': {
-			if (quantity > 500) return "I can only kill 500 Bryophyta's at a time!";
-			const loot = bryophyta.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'ALCHEMICALHYDRA':
-		case 'ALCHEMICALHYDRAS':
-		case 'AHYDRA':
-		case 'AHYDRAS':
-		case 'HYDRAS':
-		case 'HYDRA':
-		case 'HYDRABOSS': {
-			if (quantity > 100000) return 'I can only kill 100000 Alchemical Hydras at a time!';
-			const loot = alchemicalHydra.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'NORMALHYDRA':
-		case 'NORMALHYDRAS':
-		case 'BABYHYDRA':
-		case 'BABYHYDRAS': {
-			if (quantity > 500) return 'I can only kill 500 Hydras at a time!';
-			const loot = hydra.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'DRAKES':
-		case 'DRAKE': {
-			if (quantity > 500) return 'I can only kill 500 Drakes at a time!';
-			const loot = drake.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'WYRMS':
-		case 'WYRM': {
-			if (quantity > 500) return 'I can only kill 500 Wyrms at a time!';
-			const loot = wyrm.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'HESPORIS':
-		case 'HESPORI': {
-			if (quantity > 100) return 'I can only kill 100 Hesporis at a time!';
-			const loot = hespori.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'GAUNTLET': {
-			if (quantity > 500) return 'I can only kill 500 Gauntlets at a time!';
-			const loot = gauntlet.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'CORRUPTEDGAUNTLET':
-		case 'CG': {
-			if (quantity > 500) return 'I can only kill 500 Gauntlets at a time!';
-			const loot = gauntlet.kill(quantity, true);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		case 'BASKNIGHT':
-		case 'BASILISKKNIGHT':
-		case 'BK':
-		case 'BASILISK': {
-			if (quantity > 300) {
-				return 'I can only kill 300 Basilik Knights at a time!';
-			}
-			const loot = basKnight.kill(quantity);
-			return loot.length > 0 ? loot : 'You got nothing.';
-		}
-
-		default:
-			return "I don't have that monster!";
-	}
-};
