+++ package.json
@@ -20,12 +20,13 @@
 		"moment": "^2.21.0",
 		"mongodb": "^3.3.3",
 		"node-fetch": "^2.6.0",
-		"oldschooljs": "^0.8.0",
+		"oldschooljs": "^0.8.4",
 		"p-limit": "^2.2.1",
 		"rethinkdbdash": "^2.3.31",
 		"rss-parser": "^3.1.1",
 		"snoostorm": "^1.2.2",
 		"snoowrap": "^1.19.0",
+		"threads": "^1.0.0",
 		"twit": "^2.2.9",
 		"utf-8-validate": "^5.0.2",
 		"zlib-sync": "^0.1.5"
