+++ src/tasks/prices.ts
@@ -3,6 +3,7 @@ import { Items } from 'oldschooljs';
 
 export default class extends Task {
 	async init() {
+		return;
 		this.run();
 	}
